Decouple unit style cache from OpenLayers
Keep the engine-neutral style cache and invalidation API in
src/geo/unitStyles.ts; move OL Icon/Style construction to
src/geo/engines/openlayers/unitStyles.ts.
